refactor(gas-estimator): :art: reorganize directories
