+++ src/gas-estimator/ArbitrumGasEstimator.ts
@@ -1,18 +1,16 @@
 import { EVMGasEstimator } from "./EVMGasEstimator";
-import {
-  ARBITRUM_L1_FEE_GAS_PRICE_ORACLE_ABI,
-  NODE_INTERFACE_ARBITRUM_ADDRESS,
-} from "../entry-point-v6";
-import { EntryPointVersion } from "../../entrypoint/shared/types";
+import { EntryPointVersion } from "../entrypoint/shared/types";
 import { Hex } from "viem";
-import { EntryPointV6 } from "../../entrypoint/v0.6.0/EntryPointV6";
-import { EntryPointV7Simulations } from "../../entrypoint/v0.7.0/EntryPointV7Simulations";
+import { EntryPointV6 } from "../entrypoint/v0.6.0/EntryPointV6";
+import { EntryPointV7Simulations } from "../entrypoint/v0.7.0/EntryPointV7Simulations";
 import {
   isUserOperationV6,
   UserOperation,
   validateUserOperation,
 } from "./UserOperation";
 import z from "zod";
+import { NODE_INTERFACE_ARBITRUM_ADDRESS } from "./constants";
+import { ARBITRUM_L1_FEE_GAS_PRICE_ORACLE_ABI } from "./abi";
 
 export class ArbitrumGasEstimator extends EVMGasEstimator {
   override async estimatePreVerificationGas(
