+++ src/gas-estimator/constants.ts
@@ -15,3 +15,16 @@ export const defaultGasOverheads = {
 
 //https://github.com/eth-infinitism/account-abstraction/blob/6f02f5a28a20e804d0410b4b5b570dd4b076dcf9/contracts/core/EntryPoint.sol#L39
 export const INNER_GAS_OVERHEAD = 10000n;
+
+export const NODE_INTERFACE_ARBITRUM_ADDRESS =
+  "0x00000000000000000000000000000000000000C8";
+export const OPTIMISM_L1_GAS_PRICE_ORACLE_ADDRESS =
+  "0x420000000000000000000000000000000000000F";
+export const MANTLE_BVM_GAS_PRICE_ORACLE_ADDRESS =
+  "0x420000000000000000000000000000000000000F";
+export const SCROLL_L1_GAS_PRICE_ORACLE_ADDRESS =
+  "0x5300000000000000000000000000000000000002";
+export const MORPH_L1_GAS_PRICE_ORACLE_ADDRESS =
+  "0x530000000000000000000000000000000000000f";
+
+export const MANTLE_L1_ROLL_UP_FEE_DIVISION_FACTOR = 1000000n;
