+++ src/legacy/entry-point-v6/GasEstimator/KakarotGasEstimator.ts
