+++ src/legacy/entry-point-v6/constants/index.ts
@@ -17,17 +17,6 @@ export const defaultGasOverheads = {
   sigSize: 65,
 };
 
-export const NODE_INTERFACE_ARBITRUM_ADDRESS =
-  "0x00000000000000000000000000000000000000C8";
-export const OPTIMISM_L1_GAS_PRICE_ORACLE_ADDRESS =
-  "0x420000000000000000000000000000000000000F";
-export const MANTLE_BVM_GAS_PRICE_ORACLE_ADDRESS =
-  "0x420000000000000000000000000000000000000F";
-export const SCROLL_L1_GAS_PRICE_ORACLE_ADDRESS =
-  "0x5300000000000000000000000000000000000002";
-export const MORPH_L1_GAS_PRICE_ORACLE_ADDRESS =
-  "0x530000000000000000000000000000000000000f";
-
 export const INITIAL_VGL_LOWER_BOUND = 0n;
 export const INITIAL_VGL_UPPER_BOUND = 30_000_000n;
 export const VGL_ROUNDING = 1n;
