+++ src/legacy/entry-point-v6/index.ts
