+++ src/legacy/entry-point-v6/interface/index.ts
