+++ src/legacy/entry-point-v6/types/index.ts
