+++ src/legacy/entry-point-v7/index.ts
