+++ src/legacy/entry-point-v7/smoke-test/SMOKES.md
