+++ src/legacy/entry-point-v7/smoke-test/arbitrum-stack.test.ts
