+++ src/legacy/entry-point-v7/smoke-test/op-stack.test.ts
